Time: 54 ms (16.46%) | Memory: 46.7 MB (13.10%) - LeetSync

+++ leetcode/264-ugly-number-ii/ugly-number-ii.java
@@ -19,7 +19,7 @@ class Solution {
                 set.add(num*5);
             }
         }
-    int val = Math.toIntExact( heap.poll());
+    int val = (int)((long) heap.poll());
         return val;
 
     }
